FirstCommit

+++ boundingBoxProblem/DataModel/wordPropertiesDataModel.cs
@@ -4,8 +4,8 @@
     {
         public int x { get; set; }
         public int y { get; set; }
-        public int width { get; set; }
-        public int height { get; set; }
+        public double width { get; set; }
+        public double height { get; set; }
     }
